Revert "feat: implement habit view (#8)"
This reverts commit c4d4528ad5544c5d52dbad85b2a56a49f1515026.
