--- HabitPlus/Common/Extensions/View+Extension.swift
@@ -1,24 +1,0 @@
-//
-//  View+Extension.swift
-//  HabitPlus
-//
-//  Created by Lucas De Oliveira Silva Firmino on 13/11/24.
-//
-
-import SwiftUI
-
-extension View {
-    func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
-        clipShape(RoundedCorner(radius: radius, corners: corners))
-    }
-}
-
-struct RoundedCorner: Shape {
-    var radius: CGFloat = .infinity
-    var corners: UIRectCorner = .allCorners
-    
-    func path(in rect: CGRect) -> Path {
-        let path = UIBezierPath(roundedRect: rect, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
-        return Path(path.cgPath)
-    }
-}
