+++ HabitPlus/Common/Lang/String+Extension.swift
@@ -1,0 +1,16 @@
+//
+//  String+Extension.swift
+//  HabitPlus
+//
+//  Created by Lucas De Oliveira Silva Firmino on 25/10/24.
+//
+
+import Foundation
+
+extension String {
+    func isEmail() -> Bool {
+        let emailRegex = "[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,64}"
+        let emailPredicate = NSPredicate(format: "SELF MATCHES %@", emailRegex)
+        return emailPredicate.evaluate(with: self)
+    }
+}
