+++ HabitPlus/Common/Network/WebService.swift
@@ -15,8 +15,6 @@ enum WebService {
         case postUser = "/users"
         case login = "/auth/login"
         case refreshToken = "/auth/refresh-token"
-        case habits = "/users/me/habits"
-        case habitValues = "/users/me/habits/%d/values"
     }
     
     enum NetworkError {
@@ -43,15 +41,15 @@ enum WebService {
         case formaUrl = "application/x-www-form-urlencoded"
     }
     
-    private static func completeUrl(path: String) -> URLRequest? {
-        guard let url = URL(string: "\(Endpoint.base.rawValue)\(path)") else { return nil}
+    private static func completeUrl(path: Endpoint) -> URLRequest? {
+        guard let url = URL(string: "\(Endpoint.base.rawValue)\(path.rawValue)") else { return nil}
         
         return URLRequest(url: url)
     }
     
     
     private static func call(
-        path: String,
+        path: Endpoint,
         method: Method,
         contentType: ContentType,
         data: Data?,
@@ -97,39 +95,8 @@ enum WebService {
             }
     }
     
-    static func call(
-        path: Endpoint,
-        method: Method = .get,
-        completion: @escaping (Result) -> Void
-    ) {
-        call(
-            path: path.rawValue,
-            method: method,
-            contentType: .json,
-            data: nil,
-            completion: completion
-        )
-    }
-    
-    
     static func call<T: Encodable>(
         path: Endpoint,
-        method: Method = .post,
-        body: T,
-        completion: @escaping (Result) -> Void
-    ) {
-        guard let jsonData = try? JSONEncoder().encode(body) else { return }
-        call(
-            path: path.rawValue,
-            method: method,
-            contentType: .json,
-            data: jsonData,
-            completion: completion
-        )
-    }
-    
-    static func call<T: Encodable>(
-        path: String,
         method: Method = .post,
         body: T,
         completion: @escaping (Result) -> Void
@@ -150,14 +117,14 @@ enum WebService {
         params: [URLQueryItem],
         completion: @escaping (Result) -> Void
     ) {
-        guard let urlRequest = completeUrl(path: path.rawValue) else { return }
+        guard let urlRequest = completeUrl(path: path) else { return }
         guard let absolueteURL = urlRequest.url?.absoluteString else { return }
         
         var components = URLComponents(string: absolueteURL)
         components?.queryItems = params
         
         call(
-            path: path.rawValue,
+            path: path,
             method: method,
             contentType: .formaUrl,
             data: components?.query?.data(using: .utf8),
