--- HabitPlus/Habit/Model/HabitResponse.swift
@@ -1,28 +1,0 @@
-//
-//  HabitResponse.swift
-//  HabitPlus
-//
-//  Created by Lucas De Oliveira Silva Firmino on 18/11/24.
-//
-
-import Foundation
-
-struct HabitResponse: Decodable {
-    
-    let id: Int
-    let name: String
-    let label: String
-    let iconUrl: String?
-    let value: Int?
-    let lastDate: String?
-    
-    enum CodingKeys: String, CodingKey {
-        case id
-        case name
-        case label
-        case iconUrl = "icon_url"
-        case value
-        case lastDate = "last_date"
-        
-    }
-}
