--- HabitPlus/HabitDetail/ViewModel/HabitDetailViewModel.swift
@@ -1,53 +1,0 @@
-//
-//  HabitDetailViewModel.swift
-//  HabitPlus
-//
-//  Created by Lucas De Oliveira Silva Firmino on 21/11/24.
-//
-
-import Foundation
-import SwiftUI
-import Combine
-
-class HabitDetailViewModel: ObservableObject {
-    @Published var uiState: HabitDetailUIState = .none
-    @Published var value: String = ""
-    
-    private var cancellable: AnyCancellable?
-    
-    let id: Int
-    let name: String
-    let label: String
-    let interactor: HabitDetailInteractor
-    
-    init(id: Int, name: String, label: String, interactor: HabitDetailInteractor) {
-        self.id = id
-        self.name = name
-        self.label = label
-        self.interactor = interactor
-    }
-    
-    deinit {
-        cancellable?.cancel()
-    }
-    
-    func save() {
-        self.uiState = .loading
-        
-        cancellable = interactor.save(habitId: id, habitValueRequest: HabitValueRequest(value: value))
-            .receive(on: DispatchQueue.main)
-            .sink(receiveCompletion: { completion in
-                switch completion {
-                case .failure(let appError):
-                    self.uiState = .error(appError.message)
-                case .finished:
-                    break
-                }
-            }, receiveValue: { created in
-                if created {
-                    self.uiState = .success
-                }
-            })
-    }
-    
-}
