--- HabitPlus/HabitDetail/ViewModel/HabitValueRequest.swift
@@ -1,12 +1,0 @@
-//
-//  HabitValueRequest.swift
-//  HabitPlus
-//
-//  Created by Lucas De Oliveira Silva Firmino on 21/11/24.
-//
-
-import Foundation
-
-struct HabitValueRequest: Encodable {
-    let value: String
-}
