+++ HabitPlus/Home/ViewModel/HomeViewModel.swift
@@ -9,11 +9,10 @@ import SwiftUI
 
 class HomeViewModel: ObservableObject {
     
-    let viewModel = HabitViewModel(interactor: HabitInteractor())
 }
 
 extension HomeViewModel {
     func habitView() -> some View {
-        return HomeViewRouter.makeHabitView(viewModel: viewModel)
+        return HomeViewRouter.makeHabitView()
     }
 }
